fix: resolve all -Wextra sign-compare and missing-field-initializer warnings
Fix sign-compare warnings:
- tt_stl/tests/test_small_vector.cpp: cast kInlineCapacity comparisons
- tt_stl/tt_stl/reflection.hpp: use std::size_t for loop index
- ttnn/cpp/ttnn/operations/functions.hpp: fix iteration variable types
- ttnn/cpp/ttnn/operations/ccl/ccl_common.hpp: cast slice_dim to std::size_t
- tests/tt_metal/test_utils/test_common.hpp: use std::size_t for loop index
- tools/scaleout/validation/utils/cluster_validation_utils.cpp: fix uint32_t vs protobuf int comparisons
- tools/scaleout/factory_system_descriptor/utils.cpp: cast protobuf size() to uint32_t
- tools/scaleout/cabling_generator/cabling_generator.cpp: cast protobuf size() to uint32_t

Fix missing designated field initializers:
- tools/scaleout/validation/utils/cluster_validation_utils.cpp: add .defines = {} to EthernetConfig
- tools/scaleout/cabling_generator/cabling_generator.cpp: add .node_type = "" to Host
- ttnn/cpp/ttnn-nanobind/program_descriptors.cpp: add .unpack_to_dest_mode = {} to ComputeConfigDescriptor
- ttnn/cpp/ttnn-nanobind/pytensor.cpp: add .bottom_up = std::nullopt to DeviceLocalBufferConfig

Suppress -Wextra warnings in third-party TracyClient target via CMake.

+++ tt_stl/tests/test_small_vector.cpp
@@ -157,7 +157,7 @@ TEST_F(SmallVectorTrackedTest, CopyConstructorSmallAndLarge) {
     // Now create a vector that requires heap storage and copy it.
     Vec large;
     // Push more than the inline capacity to force heap allocation.
-    for (int i = 0; i < kInlineCapacity + 4; ++i) {
+    for (int i = 0; i < static_cast<int>(kInlineCapacity) + 4; ++i) {
         large.emplace_back(i);
     }
     auto* const largeData = large.data();
@@ -191,7 +191,7 @@ TEST_F(SmallVectorTrackedTest, MoveConstructorSmallAndLarge) {
     }
     {
         Vec large;
-        for (int i = 0; i < kInlineCapacity + 2; ++i) {
+        for (int i = 0; i < static_cast<int>(kInlineCapacity) + 2; ++i) {
             large.emplace_back(i);
         }
         // Force allocation beyond inline capacity
@@ -215,7 +215,7 @@ TEST_F(SmallVectorTrackedTest, MoveConstructorSmallAndLarge) {
 TEST_F(SmallVectorTrackedTest, DestructorCalls) {
     {
         Vec vec;
-        for (int i = 0; i < kInlineCapacity; ++i) {
+        for (int i = 0; i < static_cast<int>(kInlineCapacity); ++i) {
             vec.emplace_back(i);
         }
         EXPECT_EQ(Tracked::dtorCount, 0);
@@ -239,7 +239,7 @@ TEST_F(SmallVectorIntTest, PushWithinSmallBufferKeepsInplaceStorage) {
     vec.push_back(0);
     auto* initialPtr = vec.data();
     auto initialCap = vec.capacity();
-    for (int i = 1; i < kInlineCapacity; ++i) {
+    for (int i = 1; i < static_cast<int>(kInlineCapacity); ++i) {
         vec.push_back(i);
         EXPECT_EQ(vec.data(), initialPtr);
         EXPECT_EQ(vec.capacity(), initialCap);
@@ -249,7 +249,7 @@ TEST_F(SmallVectorIntTest, PushWithinSmallBufferKeepsInplaceStorage) {
 
 TEST_F(SmallVectorIntTest, PushBeyondSmallBufferTriggersHeapAllocation) {
     Vec vec;
-    for (int i = 0; i < kInlineCapacity; ++i) {
+    for (int i = 0; i < static_cast<int>(kInlineCapacity); ++i) {
         vec.push_back(i);
     }
     auto* const smallPtr = vec.data();
@@ -394,7 +394,7 @@ TEST_F(SmallVectorIntTest, SwapExchangesContentsAndStorage) {
     Vec small{1, 2, 3};
     Vec large;
     // Force large to allocate on the heap by pushing beyond inline capacity.
-    for (int i = 0; i < kInlineCapacity + 2; ++i) {
+    for (int i = 0; i < static_cast<int>(kInlineCapacity) + 2; ++i) {
         large.push_back(i + 10);
     }
     std::size_t largeCap = large.capacity();
@@ -457,7 +457,7 @@ TEST_F(SmallVectorIntTest, SizeCapacityAndEmpty) {
     EXPECT_EQ(vec.size(), 1u);
     auto smallCap = vec.capacity();
     // Fill up to small buffer capacity.
-    for (int i = 1; i < kInlineCapacity; ++i) {
+    for (int i = 1; i < static_cast<int>(kInlineCapacity); ++i) {
         vec.push_back(i + 1);
     }
     EXPECT_EQ(vec.size(), kInlineCapacity);
